Limited progress made (damn it)
The problem is getting the values of the activity to be edited to
display in the inputs.  It works - but only if you use primitive scope
variables.  On the other hand, using primitive scope variables means
that the model doesn't update when the user makes changes - but if you
use an object scope variable, then the values of the activity to be
edited do NOT appear in the input fields.

Long story short - I'm about two seconds from tearing my hair out.  Type
changing works, but only because button click events force the model to
change.